refactor: reapply reverted changes
