+++ app/build.gradle.kts
@@ -151,12 +151,12 @@ dependencies {
     implementation(libs.constraintlayout)
     implementation(libs.appcompat)
     implementation(libs.core.ktx)
+    implementation(libs.documentfile)
     implementation(libs.gridlayout)
     implementation(libs.preference.ktx)
     implementation(libs.work.runtime.ktx)
     implementation(libs.navigation.fragment.ktx)
     implementation(libs.navigation.ui.ktx)
-    implementation(libs.legacy.support.v4)
     ksp(libs.room.compiler)
     implementation(libs.room.runtime)
     implementation(libs.room.ktx)
