+++ app/src/main/java/com/kylecorry/trail_sense/tools/navigation/ui/NavigationSheetView.kt
@@ -24,7 +24,7 @@ import com.kylecorry.trail_sense.shared.DistanceUtils.toRelativeDistance
 import com.kylecorry.trail_sense.shared.FormatService
 import com.kylecorry.trail_sense.shared.Units
 import com.kylecorry.trail_sense.shared.UserPreferences
-import com.kylecorry.trail_sense.shared.extensions.NavigationSensorValues
+import com.kylecorry.trail_sense.shared.sensors.NavigationSensorValues
 import com.kylecorry.trail_sense.shared.navigateWithAnimation
 import com.kylecorry.trail_sense.shared.views.DataPointView
 import com.kylecorry.trail_sense.tools.beacons.domain.Beacon
