+++ gradle/libs.versions.toml
@@ -6,7 +6,7 @@ androidXTestUiAutomator = "2.3.0"
 androidXArchCore = "2.2.0"
 androidXCoreRemoteViews = "1.1.0"
 androidXExifinterfaceVersion = "1.4.2"
-andromedaVersion = "20.0.1"
+andromedaVersion = "21.0.0"
 cameraxVersion = "1.6.0"
 coreKtx = "1.18.0"
 appcompat = "1.7.1"
@@ -16,13 +16,13 @@ desugar_jdk_libs = "2.1.5"
 flexbox = "3.0.0"
 fragment = "1.8.9"
 gridlayout = "1.1.0"
+documentfile = "1.1.0"
 junitJupiter = "5.13.4"
 junitPlatformRunner = "1.14.4"
 junitVersion = "1.3.0"
 junit = "4.13.2"
 kotlinxCoroutinesAndroid = "1.10.2"
 ksp = "2.3.4"
-legacySupportV4 = "1.0.0"
 navVersion = "2.9.8"
 lifecycleVersion = "2.10.0"
 material = "1.14.0"
@@ -87,6 +87,7 @@ core-ktx = { module = "androidx.core:core-ktx", version.ref = "coreKtx" }
 appcompat = { module = "androidx.appcompat:appcompat", version.ref = "appcompat" }
 constraintlayout = { module = "androidx.constraintlayout:constraintlayout", version.ref = "constraintlayout" }
 desugar_jdk_libs = { module = "com.android.tools:desugar_jdk_libs", version.ref = "desugar_jdk_libs" }
+documentfile = { module = "androidx.documentfile:documentfile", version.ref = "documentfile" }
 ext-junit = { module = "androidx.test.ext:junit", version.ref = "junitVersion" }
 flexbox = { module = "com.google.android.flexbox:flexbox", version.ref = "flexbox" }
 gridlayout = { module = "androidx.gridlayout:gridlayout", version.ref = "gridlayout" }
@@ -96,7 +97,6 @@ junit-jupiter-api = { module = "org.junit.jupiter:junit-jupiter-api", version.re
 junit-platform-runner = { module = "org.junit.platform:junit-platform-runner", version.ref = "junitPlatformRunner" }
 junit = { module = "junit:junit", version.ref = "junit" }
 kotlinx-coroutines-android = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-android", version.ref = "kotlinxCoroutinesAndroid" }
-legacy-support-v4 = { module = "androidx.legacy:legacy-support-v4", version.ref = "legacySupportV4" }
 lifecycle-viewmodel-ktx = { module = "androidx.lifecycle:lifecycle-viewmodel-ktx", version.ref = "lifecycleVersion" }
 lifecycle-service = { module = "androidx.lifecycle:lifecycle-service", version.ref = "lifecycleVersion" }
 luna = { module = "com.kylecorry:luna", version.ref = "luna" }
